custom prompter with colored input text
